third_party/WebKit/LayoutTests/webaudio/osc-square-sweep-snr-expected.txt

Test Square Oscillator with Exponential Sweep

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 81.57 dB
PASS Maximum difference below threshold of 7.26 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 18260 out of 176400
PASS successfullyParsed is true

TEST COMPLETE
